Add metadata structure
